2dGame
A 2D game I made with Kotlin language. Made for Android devices. When you press and hold the screen, the character goes up and when you release it, it falls down. If you eat the oncoming shapes, you earn points, but if you eat the black one, the game is over.